trim the example
