+++ harness/determined/cli/cli.py
@@ -30,7 +30,7 @@ from determined.cli.tensorboard import args_description as tensorboard_args_desc
 from determined.cli.trial import args_description as trial_args_description
 from determined.cli.user import args_description as user_args_description
 from determined.cli.version import args_description as version_args_description
-from determined.common.api.job import args_description as job_args_description
+from determined.cli.job import args_description as job_args_description
 from determined.cli.version import check_version
 from determined.common import api, yaml
 from determined.common.api import authentication, certs
